Modify Features to extract history of test assmnts
Basically, Features now can accept a list of assessments whose history is to be extracted, rather than extracting the history all the assessments with valid attempts.

This modification was necessary because the record of submission assessments has been expunged from the test data. These assessemnts therefore do not a record of any valid attempts and hence would not have been considered by the previous version of Features.